requirements.txt
# VOT1 Dependencies

# Core dependencies
requests>=2.25.0
aiohttp>=3.7.4
asyncio>=3.4.3
psutil>=5.9.0
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.64.0
flask>=2.2.0  # For local MCP bridge
flask-cors>=3.0.10  # For local MCP bridge

# NLP and AI
openai>=1.3.0  # For OpenAI API
anthropic>=0.5.0  # For Anthropic Claude models
numpy>=1.22.0
scipy>=1.8.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.2  # For vector embeddings

# Knowledge representation
owlready2>=0.38  # For OWL reasoning engine
networkx>=2.8.0  # For graph representations
rdflib>=6.2.0  # For RDF parsing and serialization

# Vector database
faiss-cpu>=1.7.2  # Vector similarity search
chromadb>=0.4.0  # Vector database

# GitHub integration
pygithub>=1.55  # GitHub API client
composio-openai>=0.1.0  # Composio integration for GitHub and other tools

# Web and API
gunicorn>=20.1.0
pydantic>=1.10.0  # For data validation

# Dashboard UI
dash>=2.6.0
dash-bootstrap-components>=1.0.0
plotly>=5.10.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

# Development tools
black>=22.6.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.971

# For CLI tools
click>=8.1.3

# For async processing
aiofiles>=22.1.0
aiomultiprocess>=0.9.0

# For documentation
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.1

# Memory Management
pymilvus>=2.3.0
langchain>=0.0.312

# Web Server & Dashboard
flask-socketio>=5.3.4
tenacity>=8.2.3
uuid>=1.30
rich>=13.4.2
sqlitedict>=2.1.0

# Storage & Database
sqlalchemy>=2.0.20
alembic>=1.12.0

# Flask and web server
Flask>=2.2.3
Werkzeug>=2.2.3
flask-socketio>=5.3.3
eventlet>=0.33.3
gunicorn>=20.1.0

# Database and storage
redis>=4.5.4
pymongo>=4.3.3

# AI and machine learning
anthropic>=0.5.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.2
scikit-learn>=1.2.2

# Utilities
python-dotenv>=1.0.0
requests>=2.28.2
psutil>=5.9.4
tenacity>=8.2.2
aiohttp>=3.8.4
pydantic>=1.10.7

# UI and visualization
plotly>=5.13.1
dash>=2.9.1
pyvis>=0.3.2

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-flask>=1.2.0

# Documentation
mkdocs>=1.4.2
mkdocs-material>=9.1.4

# MCP integrations
jupyter>=1.0.0
python-multipart>=0.0.6

# Blockchain Integration
solana-py>=0.30.2  # Python client for Solana
web3>=6.5.0  # For Ethereum integration
pycryptodome>=3.18.0  # Cryptographic library for blockchain operations
coincurve>=18.0.0  # Required for Bitcoin and Solana address handling

# Hex Utilities
pysha3>=1.0.2  # SHA3 hash functions
base58>=2.1.1  # Base58 encoding/decoding for blockchain addresses
hexbytes>=0.3.0  # Hex string utility library

# Zero-Knowledge Proofs
zokrates-pycrypto>=0.2.2  # ZoKrates binding for Python for ZK proofs generation and verification
snarkjs>=0.7.0  # JavaScript library for zkSNARK
circom>=0.5.45  # Compiler for circuit definitions in the circom language 